implements HOF to manipulate state
Obstacle object can take the Ball as input, and modify the bucket counting due to this higher order function. Counting logic is handled in the inner function, the HOF is only allowed to take Ball object (which touching the finish line). Also adds bucket count visualization by generating a histogram as the balls fall